% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar.R
\docType{data}
\name{build_time}
\alias{build_time}
\title{leqtar}
\format{An object of class \code{character} of length 1.}
\usage{
build_time
}
\seealso{
https://github.com/ErikSchutte/leqtar/issues
}
\author{
Erik Schutte
}
\keyword{datasets}
