% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_genotypes_to_frequencies.R
\name{change_genotype_to_frequency}
\alias{change_genotype_to_frequency}
\title{change_allele_to_requencies}
\usage{
change_genotype_to_frequency(x, allele.list, VERBOSE)
}
\arguments{
\item{x}{current cell.}

\item{allele.list}{list containg the current alleles per snp.}

\item{VERBOSE}{verbosity true/false.}
}
\description{
Function to identify the genotype and determine wheter it is major, minor or heterozygote.
}
