% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_process_files.R
\name{check_object_or_file}
\alias{check_object_or_file}
\title{check_object_or_file}
\usage{
check_object_or_file(path_argument, data_argument, name_argument)
}
\arguments{
\item{path_argument}{the argument that should contain the path.}

\item{data_argument}{the argument that contains the object.}

\item{name_argument}{variable name to specify genotype phenotype covariaties or w/e in the messages.}
}
\value{
If the path_argument is indeed given return the file content, else return the object.
}
\description{
Checks wether the given parameter is a file path or an R object. Thus determines wether to read it or not.
}
