% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_trityper_to_text.R
\name{complementSNP}
\alias{complementSNP}
\title{complementSNP}
\usage{
complementSNP(allele)
}
\arguments{
\item{allele}{input allele.}
}
\value{
returns complement from allele.
}
\description{
Calculates the complement for input allele.
}
