% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariate_test_data}
\alias{covariate_test_data}
\title{Example of a covariate data set.}
\format{A data frame with 2 rows and 200 variables:
\describe{
  \item{sample1}{Sample 1 with covariates such as gender and time.}
  \item{sample2}{Sample 2 with covariates such as gender and time.}
  \item{sample3}{Sample 3 with covariates such as gender and time.}
}}
\usage{
covariate_test_data
}
\description{
This data set contains two covariates, age and gender. Gender is denoted as 0 for female and 1 for male.
}
\keyword{datasets}
