% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression_test_data}
\alias{expression_test_data}
\title{Example of an expression data set.}
\format{A data frame with 3 rows and 3 variables:
\describe{
  \item{sample1}{Sample 1 with expression in raw read counts per gene.}
  \item{sample2}{Sample 2 with expression in raw read counts per gene.}
  \item{sample3}{Sample 3 with expression in raw read counts per gene.}
}}
\usage{
expression_test_data
}
\description{
This data set contains some expression values. These values represent raw read counts.
}
\keyword{datasets}
