% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_process_files.R
\name{filter_variants_by_groups}
\alias{filter_variants_by_groups}
\title{filter_variants_by_groups}
\usage{
filter_variants_by_groups(genotype_file_content)
}
\arguments{
\item{genotype_file_content}{the genotype data.}
}
\value{
filtered genotype data.
}
\description{
Filters variants by genotype groups. A minimum of two groups have to exsist for a variant
to be considered for analysis.
}
