% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode_names}
\alias{gencode_names}
\title{Gencode gene id's and gene names.}
\format{A data frame with 60252 rows and 2 variables
\describe{
  \item{gene_id}{Gencode gene id's.}
  \item{gene_name}{Corrosponding gene name.}
}}
\usage{
gencode_names
}
\description{
This data set contains a column with gencode gene ids and gencode gene names.
}
\keyword{datasets}
