% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gencode_position}
\alias{gencode_position}
\title{Gene positions.}
\format{A data frame with 60252 rows ans 4 variables
\describe{
  \item{geneid}{The Id's.}
  \item{chr}{located on chromosome .. .}
  \item{s1}{start position of the gene.}
  \item{s2}{stop position of the gene.}
}}
\usage{
gencode_position
}
\description{
This data set contain the positions for each gene id on the genome.
}
\keyword{datasets}
