% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genotype_test_data}
\alias{genotype_test_data}
\title{Example of a genotype data set.}
\format{A data frame with 200 rows and 500 variables:
\describe{
  \item{rs000001}{Example SNP with rs number 1}
  \item{rs000002}{Example SNP with rs number 2}
  \item{rs000003}{Example SNP with rs number 3}
  ...
}}
\usage{
genotype_test_data
}
\description{
This data set contains some genotypes.
}
\keyword{datasets}
