% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genotype_test_locations}
\alias{genotype_test_locations}
\title{Example of a genotype position data set.}
\format{A data frame with 200 rows and 3 variables:
\describe{
  \item{snp}{SNP IDs.}
  \item{chr}{The chr location for each SNP.}
  \item{pos}{The position for each SNP.}
  ...
}}
\usage{
genotype_test_locations
}
\description{
This data set contains some genotypes.
}
\keyword{datasets}
