% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gepo}
\alias{gepo}
\title{Genotype positions.}
\format{A data frame with 5493617 rows and 3 variables.
\describe{
  \item{snp}{The rs id for SNPs.}
  \item{chr}{The chromosome location.}
  \item{pos}{The position on the genomte.}
}}
\usage{
gepo
}
\description{
This data set contains the positions for each SNP.
}
\keyword{datasets}
