% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar.R
\name{init_basic_directories}
\alias{init_basic_directories}
\title{Creates initial file structure for leqtar output.}
\usage{
init_basic_directories(arguments)
}
\arguments{
\item{arguments}{the empty/partially empty arguments variable used for processing.}
}
\value{
arguments the partially empty arguments variables used for processing.
}
\description{
Creates initial file structure for leqtar output.
}
