% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_trityper_to_text.R
\name{inverseDosage}
\alias{inverseDosage}
\title{inverseDosage}
\usage{
inverseDosage(dosageVector)
}
\arguments{
\item{dosageVector}{vector of dosage values per rsID.}
}
\value{
inverse dosage vector.
}
\description{
inverts dosage according to vector.
}
