% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_analysis.R
\name{leqtar_analysis}
\alias{leqtar_analysis}
\title{leqtar_analysis function}
\usage{
leqtar_analysis(dataFiles, arguments)
}
\arguments{
\item{dataFiles}{a named list containing the data files that were processed by leqtar_process_files.R}

\item{arguments}{parsed and processed arguments}
}
\description{
Uses linear regression analysis to identify the influence of a specific genotype on expression.
}
