% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_create_manhattan_plots.R
\name{leqtar_create_manhattan_plots}
\alias{leqtar_create_manhattan_plots}
\title{leqtar_create_manhattan_plots}
\usage{
leqtar_create_manhattan_plots()
}
\description{
Creates manhattan plots from the data file produced by leqtar_analysis.
}
