% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_create_tables.R
\name{leqtar_create_tables}
\alias{leqtar_create_tables}
\title{leqtar_create_tables}
\usage{
leqtar_create_tables()
}
\arguments{
\item{somePath}{change this later.}
}
\description{
Creates result tables from the data file produced by leqtar_analysis.
}
