% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_results.R
\name{leqtar_results}
\alias{leqtar_results}
\title{leqtar_results}
\usage{
leqtar_results(arguments)
}
\arguments{
\item{arguments}{contains the input arguments.}
}
\description{
Processes the results that are generated by the wrapper leqtar_analysis.
Calls helper functions to process all result files.
}
