% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_trityper_to_text.R
\name{leqtar_trityper_to_text}
\alias{leqtar_trityper_to_text}
\title{leqtar_trityper_to_text}
\usage{
leqtar_trityper_to_text(xmx = 30, xms = 30, cohort, snp_location,
  trityper_location, sample_location, output_location)
}
\arguments{
\item{xmx}{Java Xmx size in gb.}

\item{xms}{Java Xms size in gb.}

\item{cohort}{The name of the cohort for conversion.}

\item{snp_location}{A path to file that contains SNPs that will be included in the conversion. Usually this is in path/to/trityper/chr_i/SNPs.txt.}

\item{trityper_location}{A path to the trityper files.}

\item{sample_location}{A path to a file containing included samples. Useally this is in path/to/trityper/chr_i/Individuals.txt.}

\item{output_location}{A path to an output directory.}
}
\description{
Conversst Trityper output to a human readable dosage matrix.
}
\note{
Only works on Linux/MacOSX
}
