% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_trityper_to_text.R
\name{loadGenotypeData}
\alias{loadGenotypeData}
\title{loadGenotypeData}
\usage{
loadGenotypeData(basePath, dataType, cacheSize = 1000,
  variantFilter = .jnull(class = file.path("org", "molgenis", "genotype",
  "variantFilter", "VariantFilter", fsep = .Platform$file.sep)),
  sampleFilter = j.null(class = file.path("org", "molgenis", "genotype",
  "sampleFilter", "SampleFilter", fsep = .Platform$file.sep)))
}
\arguments{
\item{basePath}{basePath, i.e. often working directory of script or executable.}

\item{dataType}{TRITYPER, VCFFOLDER. Conversion for GenotypeHarmonizer | Genotype-IO}

\item{cacheSize}{Cache size.}
}
\value{
returns loaded genotype data.
}
\description{
Loads genotype data using GenotypeHarmonizer.
}
