% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotype_test_locations}
\alias{phenotype_test_locations}
\title{Example of a phenotype data set.}
\format{A data frame with 5000 rows and 4 variables:
\describe{
  \item{geneid}{Gene IDs.}
  \item{chr}{The chr location for each gene.}
  \item{s1}{The start position for each gene.}
  \item{s2}{The stop position for each gene.}
}}
\usage{
phenotype_test_locations
}
\description{
This data set contains some expression values. These values represent raw read counts.
}
\keyword{datasets}
