% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_create_genotype_boxplots.R
\name{prepare_df_qtls}
\alias{prepare_df_qtls}
\title{prepare_df_qtls}
\usage{
prepare_df_qtls(qtl, arguments)
}
\arguments{
\item{qtl}{a single qtl from matrix qtl.}

\item{arguments}{the processed input arguments.}
}
\description{
prepares the matrix qtl output data.frame and transforms it slightly. Adding information
to easily create genotype plots.
}
