% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar.R
\name{process_arguments}
\alias{process_arguments}
\title{Processes the user input arguments}
\usage{
process_arguments(genotypeFile, expressionFile, covariateFile, output_dir,
  run_name, genoToFreq, forceRun)
}
\arguments{
\item{genotypeFile}{the genotype file}

\item{expressionFile}{the expression file}

\item{covariateFile}{the covariates, either zero or defined}

\item{output_dir}{the output dir, either default or user specified}

\item{genoToFreq}{a flag that converts genotypes AA AT TT to frequencies 0 1 2}
}
\value{
arguments object, containing all processed arguments
}
\description{
Processes the user input arguments
}
