% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_process_files.R
\name{read_files}
\alias{read_files}
\title{read_files function}
\usage{
read_files(file_extension, file_path)
}
\arguments{
\item{file_extension}{the file extension}

\item{file_path}{the file path}
}
\value{
the file content
}
\description{
Tries to read in files determined by their file extensions.
Always returns the content of the given file.
}
