% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar.R
\name{init_basic_directories}
\alias{init_basic_directories}
\title{Creates initial file structure for leqtar output.}
\usage{
init_basic_directories(output_dir, run_name, arguments, forceRun)
}
\arguments{
\item{output_dir}{the output_dir variable set by the user.}

\item{run_name}{variable name for current run.}

\item{arguments}{the empty/partially empty arguments variable used for processing.}
}
\value{
output_dir contains the output path for the current run

arguments modified arguments list.
}
\description{
Creates initial file structure for leqtar output.
}
