% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar.R
\name{leqtar}
\alias{leqtar}
\title{Leqtar}
\usage{
leqtar(genotypeFile = NULL, phenotypeFile = NULL, covariateFile = NULL,
  output_dir = NULL, run_name = NULL, genoToFreq = F, forceRun = F)
}
\arguments{
\item{genotypeFile}{[REQUIRED] A Genotype file. SNP names are expected as colnames and sample names as rownames.}

\item{phenotypeFile}{[REQUIRED] A Phenotype file. Sample names are expected as colnames and stimulations/genes as rownames.}

\item{covariateFile}{[OPTIONAL] A Covariate file containing covariates. Sample names are expected as colnames and covariates as rownames.}

\item{output_dir}{[OPTIONAL] A relative path from your current working directory or an absolute path to a specific directory. The results will be stored in this location.}

\item{run_name}{[REQUIRED] A name that will be used for the current 'run'.}

\item{genoToFreq}{[OPTIONAL] Turns on conversion from genotypes i.e. 'AC' to a frequency for linear regression analysis.}

\item{forceRun}{[OPTIONAL] Normally Leqtar perserves data, by turning this to `TRUE` runs that already exist will be overwritten.}

\item{genotypePositionFile}{[OPTIONAL] A file containing the positions for each SNP. The colnames are 'snp', 'chr' and    'pos'.}

\item{phenotypePositionFile}{[OPTIONAl] A file containing the positions for each gene.}
}
\description{
Leqtar uses a linear regression models for eQTL analysis. It has a multitude of options, play around with the test data that is availble to get a good idea of your optimal settings.
Leqtar tries to do much of the normal pre-processing work for you. Leqtar corrects sample differences, dimensions problems, 'cell' type differences and other things.
Please note to always double check your results and if there are any issues, please refer to https://github.com/ErikSchutte/leqtar/issues. This is a work in progress, I expect there to be bugs.
}
\note{
For a complete view of how to run and use Leqtar, please visit https://github.com/ErikSchutte/leqtar.
}
