% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_create_genotype_boxplots.R
\name{leqtar_create_genotype_boxplots}
\alias{leqtar_create_genotype_boxplots}
\title{leqtar_create_genotype_boxplots}
\usage{
leqtar_create_genotype_boxplots(arguments)
}
\arguments{
\item{arguments}{the processed input arguments.}
}
\description{
Using the output from leqtar_analysis this function will try to create genotype boxplots for the top x associations found by Matrix eQTL.
}
