% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_create_tables.R
\name{leqtar_create_tables}
\alias{leqtar_create_tables}
\title{leqtar_create_tables}
\usage{
leqtar_create_tables(arguments, output_data, output_tbl)
}
\arguments{
\item{arguments}{The processed input arguments}

\item{output_data}{the folder containing the output from leqtar_analysis.}

\item{output_tbl}{the folder where the data tables should go.}
}
\description{
Creates result tables from the data file produced by leqtar_analysis.
}
