% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar_process_files.R
\name{leqtar_process_files}
\alias{leqtar_process_files}
\title{leqtar_process_files function}
\usage{
leqtar_process_files(arguments)
}
\arguments{
\item{arguments}{path to file supplied by user.}
}
\value{
content of data files.
}
\description{
leqtar_process_files function
}
