% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqtar.R
\name{process_arguments}
\alias{process_arguments}
\title{Processes the user input arguments}
\usage{
process_arguments(genotypeFile, expressionFile, output_dir, covariateFile,
  geneToFreq)
}
\arguments{
\item{genotypeFile}{the genotype file}

\item{expressionFile}{the expression file}

\item{output_dir}{the output dir, either default or user specified}

\item{covariateFile}{the covariates, either zero or defined}
}
\value{
arguments object, containing all processed arguments
}
\description{
Processes the user input arguments
}
